first step of cleaning up integration tests: sort into files
